requirements.txt
oandapyV20==0.7.2
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
yfinance>=0.2.40
matplotlib>=3.8.0
backtesting>=0.3.3
seaborn>=0.13.0
